requirements.txt
# PyTorch-related
torch==2.10.0
torchvision==0.25.0
tensordict==0.11.0
torchrl==0.11.1

# OGBench-based environments
ogbench==1.2.1

# Push-Tetris environment
pygame==2.6.1
pymunk==6.8.0
shapely==2.1.2

# Utils
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
termcolor==3.3.0
wandb==0.25.0
moviepy==2.2.1
opencv-python==4.13.0.92
einops==0.8.2

# Visual encoder training
matplotlib==3.10.8
accelerate==0.19.0  # for multi-GPU DLP training
